+++ lib/store_model/configuration.rb
@@ -15,6 +15,10 @@ module StoreModel
     # @return [Boolean]
     attr_accessor :serialize_unknown_attributes
 
+    # Controls if the result of `as_json` will serialize enum fiels using `as_json`
+    # @return [Boolean]
+    attr_accessor :serialize_enums_using_as_json
+
     def initialize
       @serialize_unknown_attributes = true
     end